[INTERNAL][E] Update undo logic in BaseResourceSelectionComposite
Updates the undo logic in BaseResourceSelectionComposite to work with
the new logic.

Adjusts it to no longer track grayed resources as this option is no
longer used.

Restricts the tracking to the base resources. Additional selections will
be made after the base state has been restored.

Adjusts the logic to also correctly restore the held list of selected
base resources.

Replaces the usage of Stack with LinkedList (as a Deque) to avoid the
performance penalty of using a synchronized collection.