[UI][E] Rework ProjectOptionComposite
Reworks the project option composite to work with the new resource
model. To keep the changes as small as possible, the new handling is not
yet used. Instead, the old methods are marked as deprecated. Their usage
will be replaced in a separate commit reworking the EnterProjectNamePage
(the class containing all ProjectOptionComposites for all shared
reference points).

Adjusts the composite to have three option for the user to choose:
  * create a new project
  * create a new directory
  * use an existing directory

Adds a unified handling to select a specific option programmatically and
disable all other fields.

Adds separate methods to set a specific option/input of the dialog.

Adjusts the "browse" dialog to actually return the path of the chosen
resource instead of always returning the base project for the resource.
Replaces the text of the button with the Eclipse folder icon.

Adjusts the text shown in the composite to match the new content.

Adds ReferencePointOptionResult as a data holder class for the result of
the page. This makes it easier to pass the selected option to other
classes.