[INTERNAL][E] Add methods to obtain container handle for selection
Adds methods to obtain the container handle for the selection
represented by the reference point option result.

Adds IllegalInputException. The exception is used to signal that the
input represented by the result is not valid. The reason why the result
is invalid is passed as the exception message.

The added error messages will be displayed as an error in
EnterProjectNamePage, so it is important that they are concise and also
mention which reference point they belong to.