[INTERNAL][CORE] Add sanity check preventing nested reference points
Adds a sanity check to the logic adding reference points in SarosSession
to prevent nested reference points from being created. If adding a
reference point would create such nested reference points, an error is
logged and the session is ended.

This is only meant as a sanity. Normally, this should already be
prevented by the IDE-specific UI implementation.

The user is not informed about the specific reason if the session ends
this way as the core does not have any way of showing used messages.

Nested reference points are detected using
IReferencePointComparator.areNested(...).