[INTERNAL] Move resource negotiation components to session context
Moves components only required by the resource negotiation to the
session context, in particular the IWorkspace and IChecksumCache
implementations.

Adjusts the logic using the IChecksumCache to request it from the
session context instead of from the plugin context.

Adjusts the initialization of the Eclipse checksum cache implementation
to be completely done by the PicoContainer. This is now possible as the
cyclic dependency issue with ISarosSessionManager has been resolved.

Modifies FileContentNotifierBridge to implement Startable and move the
initialization of the Eclipse workspace listener into the start block.
Also add the logic to de-register the listener in the stop block. This
was done to avoid issues with the Eclipse SarosSessionTest checking
whether any workspace listeners were present before the session was
started. As the listener in question was added when the context was
initialized and not when it was started.