[STF] Mark WriteAccessChangeAndImmediateWriteTest as flaky
Marks the test in WriteAccessChangeAndImmediateWriteTest as flaky as it
fails on the build server without an apparent reason. I could not find
out why this happens by looking at the logs or the screenshots of the
failed runs. Furthermore, the issue could not be reproduced when running
the test on its own using Eclipse (on Linux or Windows) or the docker
image.